Refactor code structure for improved readability and maintainability, updated Workflows with new naming
