+++ js/recipe_renderer.js
@@ -1,10 +1,10 @@
 /**
- * Workflow Renderer — Set initial node size to match Workflow Builder width.
+ * Recipe Renderer — Set initial node size to match Recipe Builder width.
  */
 import { app } from "../../scripts/app.js";
 
 app.registerExtension({
-    name: "prompt-manager.workflow-renderer",
+    name: "prompt-manager.recipe-renderer",
     async nodeCreated(node) {
         if (node.comfyClass !== "RecipeRenderer") return;
         requestAnimationFrame(() => {
