+++ nodes/recipe_model_loader.py
